Add more details to GUI

+++ Firmware/sub/Encoders/Encoders.h
@@ -11,6 +11,6 @@ namespace Encoders
   uint16_t getPositionSPI(uint8_t, uint8_t);
   uint16_t updateSingle(uint8_t);
   void update();
-  float getStatus(uint8_t);
+  int getStatus(uint8_t);
   uint8_t spiWriteRead(uint8_t, uint8_t, uint8_t);
 }
